+++ src/texture.h
@@ -21,7 +21,7 @@ void texture_bind(struct Texture texture);
 void texture_destroy(struct Texture* texture);
 
 struct TextureAtlas texture_atlas_create(const char* path, int tile_size);
-vec2s texture_atlas_get_uv_offset(struct TextureAtlas* texture_atlas, uint16_t tile_index);
+vec2s texture_atlas_get_uv_offset(struct TextureAtlas* texture_atlas, uint8_t tile_index);
 void texture_atlas_destroy(struct TextureAtlas* texture_atlas);
 
 #endif
